Harden Graph SLAM architecture boundaries
